Share performAnalsysis
- remove CancellableCompleter class
- Remove DelayedTimer class
