+++ pubspec.yaml
@@ -23,7 +23,7 @@ dependencies:
   yaml: ^2.2.0
   checked_yaml: ^1.0.0
   js: ^0.6.0
-  stream_transform: any
+  stream_transform: ^1.2.0
 
 dev_dependencies:
   build_runner: any
